refactor: MVVM pattern, main viewmodel main viewmodel 사용 위해 기존 getUserResponse 수정 UserInfo class 생성해 LiveData로 사용
